Combined the batch norm and batch stand into norm and stand
