--- Imaginet.Units.Math/ElementWise/BatchStandardization/BatchStandardization.imunit
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Imaginet>
-	<Unit name="Imaginet.Units.Math.BatchStandardization">
-		<DisplayName>Batch Standardization</DisplayName>
-		<DisplayPath>/Math/Element Wise</DisplayPath>
-
-		<Description>
-			<Header>Description</Header>
-			Standardize all elements in a batch tensor using the global mean and standard deviation computed across the entire batch: output = (x - mean(batch)) / std(batch).
-
-			Unlike the Standardization unit, this unit computes the mean and standard deviation at runtime from the batch tensor itself rather than using fixed design-time constants. A single mean and standard deviation is derived from all elements across all samples, so every sample is shifted and scaled relative to the same global statistics. If the entire batch is constant (standard deviation is zero), the denominator is clamped to 1 to avoid division by zero. The tensor shape and dimensions are preserved.
-
-			The input must be at least 2D with shape [num_samples, sample_size]. Supports float32 data type.
-
-			<Header>Usage</Header>
-			Use the Batch Standardization unit when you need to center and scale a collection of samples to zero mean and unit variance relative to the observed statistics of the batch, without prior knowledge of the data distribution.
-		</Description>
-
-		<Parameters>
-			<InputSocket name="input" text="Input" description="Input batch tensor to be standardized. Must be at least 2D with shape [num_samples, sample_size]. Supports float32 data type only." />
-			<Expression name="count" value="input.shape.flat" description="Total number of elements in the batch tensor (computed from flattened shape)." />
-			<OutputSocket name="output" type="input.type" shape="input.shape" text="Output" description="Output batch tensor standardized using the global mean and standard deviation of the batch. Has the same shape and data type as the input." />
-		</Parameters>
-
-		<Contracts>
-			<Assert
-				test="input.type == System.Float32"
-				error="The input tensor ({input.type}) must have type: Float32" />
-			<Assert
-				test="input.shape.count >= 2"
-				error="The input tensor must be at least 2D. Got {input.shape.count}D." />
-		</Contracts>
-
-		<Implementations>
-			<Implementation language="C" fragment="batch_standardization.h:batch_standardization_f32" call="batch_standardization_f32(input, count, output)">
-				<Conditional value="input.type == System.Float32" />
-			</Implementation>
-		</Implementations>
-
-	</Unit>
-
-</Imaginet>
